% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CasPhy_functions.R
\name{as.ultrametric}
\alias{as.ultrametric}
\title{turn the phylogeny into ultrametric phylogeny.}
\usage{
as.ultrametric(phy)
}
\arguments{
\item{phy}{The phylogeny object.}
}
\value{
an ultrametric phylogeny.
}
\description{
turn the phylogeny into ultrametric phylogeny.
}
\examples{
\dontrun{
tre<-as.ultrametric(tre)
}

}
