% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CasPhy_functions.R
\name{cal.bs}
\alias{cal.bs}
\title{calculate the bootstrap values of the phylogeny.}
\usage{
cal.bs(phy, bsphy, ncore = 1)
}
\arguments{
\item{phy}{The phylogeny object.}

\item{bsphy}{List of bootstrap trees.}

\item{ncore}{The number of threads used.}
}
\value{
An integer vector containing the bootstrap value of nodes in the phylogeny.
}
\description{
calculate the bootstrap values of the phylogeny.
}
\examples{
\dontrun{
bs_value<-cal.bs(tre,bstrees,10)
}

}
