% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CasPhy_functions.R
\name{cal.nevo}
\alias{cal.nevo}
\title{calculate the number of evolution events of the phylogeny.}
\usage{
cal.nevo(phy, mat, weight = NULL, verbose = T)
}
\arguments{
\item{phy}{The phylogeny object.}

\item{mat}{The mutation matrix.}

\item{weight}{Weights of indels.}

\item{verbose}{Whether to print the number of evolution events.}
}
\value{
an bit-coded matrix marking the evolution events in the nodes of the phylogeny.
}
\description{
calculate the number of evolution events of the phylogeny.
}
\examples{
\dontrun{
nEvo<-cal.nevo(tre,mat)
}

}
