% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CasPhy_functions.R
\name{collapse.infoless}
\alias{collapse.infoless}
\title{add the length of edges to the phylogeny according the number of evolution events and collapse nodes without evolution.}
\usage{
collapse.infoless(phy, nEvo, collapse = T)
}
\arguments{
\item{phy}{The phylogeny object.}

\item{nEvo}{The results from `cal.nevo`.}

\item{collapse}{Whether to collapse the nodes. If set to False, the function just add the length of edges to the phylogeny.}
}
\value{
a phylogeny.
}
\description{
add the length of edges to the phylogeny according the number of evolution events and collapse nodes without evolution.
}
\examples{
\dontrun{
tre<-collapse.infoless(tre,nEvo)
}

}
