% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CasPhy_functions.R
\name{h.partial.join.parallel}
\alias{h.partial.join.parallel}
\title{reconstruct the phylogenies using partial join algorithm in parallel.}
\usage{
h.partial.join.parallel(
  mat,
  splt,
  keep_cutoff = 1,
  weight = NULL,
  noise = F,
  bk_size = 40000,
  ncore = 2
)
}
\arguments{
\item{mat}{The mutation matrix.}

\item{splt}{Split of data from `split.data`.}

\item{keep_cutoff}{The cutoff to break problematic clades.}

\item{weight}{Weights of indels.}

\item{noise}{Whether add some noise to the weights to prevent equal number of shared mutations.}

\item{bk_size}{If the number of OTUs is larger than this value, the function will create a backup file to save memory.}

\item{nocre}{The number of threads used.}
}
\value{
a list of phylogenies.
}
\description{
reconstruct the phylogenies using partial join algorithm in parallel.
}
\examples{
\dontrun{
phys<-h.partial.join.parallel(imputed,splt,noise=F,ncore=10)
}

}
