% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CasPhy_functions.R
\name{insert.dup}
\alias{insert.dup}
\title{insert duplicated OTUs back to the phylogeny}
\usage{
insert.dup(phy, dup_idx, label = paste0("C", 1:length(dup_idx)))
}
\arguments{
\item{phy}{The phylogeny object.}

\item{dup_idx}{The indices of duplicated OTUs from `find.dup`.}

\item{label}{Tip label of the full phylogeny.}
}
\value{
a phylogeny.
}
\description{
insert duplicated OTUs back to the phylogeny
}
\examples{
\dontrun{
tre<-insert.dup(tre,dup_idx)
}

}
