% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CasPhy_functions.R
\name{iter.impute.packed}
\alias{iter.impute.packed}
\title{perform iterative imputation to fill the missing entries.}
\usage{
iter.impute.packed(
  mat,
  weight = NULL,
  n_neighbor = 10:2,
  p_mutation = 0.5,
  n_iter = 100,
  cutoff = max(floor(nrow(mat)/100), 100),
  verbose = F,
  ncore = 1
)
}
\arguments{
\item{mat}{The mutation matrix.}

\item{weight}{Weights of indels.}

\item{n_neighbor}{The number of neighbors used.}

\item{p_mutation}{The minimum percentage of shared mutations for neighbors.}

\item{n_iter}{The maximum number of iterations.}

\item{cutoff}{If the number of newly imputed entries is smaller than this value, the function will change the number of neighbors according to `n_neighbor`.}

\item{ncore}{The number of threads used.}

\item{verobse}{Whether to print detailed information.}
}
\value{
an imputed matrix.
}
\description{
perform iterative imputation to fill the missing entries.
}
\examples{
\dontrun{
imputed<-iter.impute.packed(mat)
}

}
