% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CasPhy_functions.R
\name{layer.summarize}
\alias{layer.summarize}
\title{summarize the results from `partial.join.DNA`}
\usage{
layer.summarize(partial_res_collected, clustIdx, splt, layer)
}
\arguments{
\item{partial_res_collected}{The results from `partial.join.DNA`.}

\item{clustIdx}{The indices of clusters.}

\item{splt}{The splt object used in phylogenetic reconstruction.}

\item{layer}{The layer of clusters.}
}
\value{
summarized results.
}
\description{
summarize the results from `partial.join.DNA`
}
\examples{
\dontrun{
mat<-read.bin2bigmat(file="test.bin")
}

}
