% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CasPhy_functions.R
\name{mat2bit}
\alias{mat2bit}
\title{Turns the matrix into bit-coded data.}
\usage{
mat2bit(mat, code_unmutated = F, return.bit = T)
}
\arguments{
\item{mat}{The mutation matrix.}

\item{code_unmutated}{Whether to code unmutated state 0.}

\item{return.bit}{Whether to return the bit-coded matrix.}
}
\value{
A list with the following components:

- Binary: is the bit-coded data.

- loci: is the offset of each locus.

- char: is the symbols in the original mutation matrix.
}
\description{
Turns the matrix into bit-coded data.
}
\examples{
\dontrun{
binary<-mat2bit(mat)
}

}
