% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CasPhy_functions.R
\name{order.tip}
\alias{order.tip}
\title{re-order the tip of the phylogeny.}
\usage{
order.tip(phy, odr)
}
\arguments{
\item{phy}{The phylogeny object.}

\item{odr}{a character vector specifies the order of tips}
}
\value{
a phylogeny.
}
\description{
re-order the tip of the phylogeny.
}
\examples{
\dontrun{
tre<-order.tip(tre,rev(tre$tip.label))
}

}
