% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CasPhy_functions.R
\name{partial.join.DNA}
\alias{partial.join.DNA}
\title{reconstruct cluster-level phylogenies}
\usage{
partial.join.DNA(
  j,
  mat_desc,
  tipname,
  matfile,
  ansMat_desc,
  nodename,
  splt,
  layer,
  prefix,
  cuda_path,
  sample_size = 0.1,
  pct_threshold = 0.001,
  nbp_threshold = 200,
  vft_thread = 8
)
}
\arguments{
\item{j}{The index of cluster.}

\item{mat_desc}{The descriptor of matrix.}

\item{tipname}{The names of tips.}

\item{matfile}{The file name of backing matrix in hard disk.}

\item{ansMat_desc}{The descriptor of matrix of pseudo-OTUs.}

\item{nodename}{The name of pseudo-OTUs.}

\item{splt}{Split of data from `split.data.DNA`.}

\item{layer}{The layer of clusters.}

\item{prefix}{The prefix of temporary file.}

\item{cuda_path}{The path of CUDA script. If set to `NULL`, do not use CUDA.}

\item{sample_size}{The percentage of outgroups sampled.}

\item{pct_threshold}{The frequency threshold used in breaking edges.}

\item{nbp_threshold}{The threshold used to filter out distal related outgroups.}

\item{vft_thread}{The number of threads used in VeryFastTree.}
}
\value{
a list containing the result.
}
\description{
reconstruct cluster-level phylogenies
}
\examples{
\dontrun{
res<-partial.join.DNA(
       mat_desc=mat_desc,
       tipname=tipname,
       matfile="test.bin",
       ansMat_desc=NULL,
       nodename=NULL,
       splt=splt,
       layer=2,
       prefix="test",
       cuda_path="~/CasPhy-cuda/build/"
     )
}

}
