% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CasPhy_functions.R
\name{phy.impute}
\alias{phy.impute}
\title{impute matrix according to a phylogeny.}
\usage{
phy.impute(mat, phy = NULL, return.ans = F)
}
\arguments{
\item{mat}{The mutation matrix.}

\item{phy}{The phylogeny.}

\item{return.ans}{Whether return the ancestral states of nodes of the phylogeny.}
}
\value{
an ultrametric phylogeny.
}
\description{
impute matrix according to a phylogeny.
}
\examples{
\dontrun{
imputed<-phy.impute(mat,tre)
}

}
