% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CasPhy_functions.R
\name{read.bin2bigmat}
\alias{read.bin2bigmat}
\title{read packed matrix from bin file into a big matrix}
\usage{
read.bin2bigmat(
  file,
  chunk_size = 32768L,
  backingfile = NULL,
  backingpath = NULL,
  descriptorfile = NULL
)
}
\arguments{
\item{file}{The file name.}

\item{chunk_size}{The chunk size used in reading.}

\item{backingfile}{The file name of backing file. If NULL, create a big matrix in memory.}

\item{backingpath}{The path to the backing file.}

\item{descriptorfile}{The file name of descriptor file.}
}
\value{
a big matrix.
}
\description{
read packed matrix from bin file into a big matrix
}
\examples{
\dontrun{
mat<-read.bin2bigmat(file="test.bin")
}

}
