% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CasPhy_functions.R
\name{read.bin2mat}
\alias{read.bin2mat}
\title{read packed matrix from bin file}
\usage{
read.bin2mat(file, chunk_size = 32768L)
}
\arguments{
\item{file}{The file name.}

\item{chunk_size}{The chunk size used in reading.}
}
\value{
an bit coded matrix containing the mutation profiles.
}
\description{
read packed matrix from bin file
}
\examples{
\dontrun{
mat<-read.bin2mat(file="test.bin")
}

}
