% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CasPhy_functions.R
\name{regularize.mat}
\alias{regularize.mat}
\title{relabel the indels.}
\usage{
regularize.mat(mat, ref)
}
\arguments{
\item{mat}{The input matrix.}

\item{ref}{the reference matrix containg all indels.}
}
\value{
an integer matrix.
}
\description{
relabel the indels.
}
\examples{
\dontrun{
mat<-regularize.mat(mat,mat)
}

}
