% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CasPhy_functions.R
\name{shared.mutation.join}
\alias{shared.mutation.join}
\title{phylogeny reconstruction using shared-mutation-join algorithm}
\usage{
shared.mutation.join(mat, weight = NULL, noise = F)
}
\arguments{
\item{mat}{The mutation matrix.}

\item{weight}{Weights of indels.}

\item{noise}{Whether add some noise to the weights to prevent equal number of shared mutations.}
}
\value{
a reconstructed phylogeny.
}
\description{
phylogeny reconstruction using shared-mutation-join algorithm
}
\examples{
\dontrun{
tre<-shared.mutation.join(mat)
}

}
