% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CasPhy_functions.R
\name{split.data.DNA}
\alias{split.data.DNA}
\title{The top solver that splits the DNA sequence data into small clusters}
\usage{
\method{split}{data.DNA}(
  mat,
  splt = NULL,
  splt_idx = 1,
  radius = 100,
  p_shared = 0.8,
  min_size = 10,
  file_backed = NULL,
  ncore = 1,
  cuda_path = NULL,
  verbose = F
)
}
\arguments{
\item{mat}{The mutation matrix.}

\item{splt}{The result from previous run.}

\item{radius}{The minimum number of mutations between neighbors.}

\item{p_shared}{The threshold for the percentage of shared neighbors.}

\item{min_size}{The minimum size of clusters. Clusters smaller than this value will be collapsed.}

\item{file_backed}{The name of backup file. If set to `NULL`, do not create backup file. If set to `""`, create an anonymous backup file.}

\item{ncore}{The number of threads used.}

\item{cuda_path}{The path of CUDA script. If set to `NULL`, do not use CUDA.}

\item{verbose}{Whether print detailed information.}

\item{split_idx}{The index of cluster to further split.}
}
\value{
Return an integer matrix containing the cluster information.
}
\description{
The top solver that splits the DNA sequence data into small clusters
}
\examples{
\dontrun{
splt<-split.data.DNA(mat,radius=1000,p_shared=0.8)
}

}
