% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CasPhy_functions.R
\name{split.data}
\alias{split.data}
\title{The top solver that splits the data into small clusters}
\usage{
\method{split}{data}(
  mat,
  weight = NULL,
  splt = NULL,
  split_idx = 1,
  radius = 0,
  p_shared = 0.8,
  min_size = 10,
  batch_size = 10000,
  file_backed = NULL,
  ncore = 1,
  verbose = F
)
}
\arguments{
\item{mat}{The mutation matrix.}

\item{weight}{Weights of indels.}

\item{splt}{The results of `split.data` from previous run.}

\item{split_idx}{The index of cluster to further split.}

\item{radius}{The minimum number of mutations between neighbors.}

\item{p_shared}{The threshold for the percentage of shared neighbors.}

\item{min_size}{The minimum size of clusters. Clusters smaller than this value will be collapsed.}

\item{batch_size}{The batch size during processing. Batch size has no effect on the results.}

\item{file_backed}{The name of backup file. If set to `NULL`, do not create backup file. If set to `""`, create an anonymous backup file.}

\item{ncore}{The number of threads used.}

\item{verbose}{Whether print detailed information.}
}
\value{
Return an integer matrix containing the cluster information.
}
\description{
The top solver that splits the data into small clusters
}
\examples{
\dontrun{
splt<-split.data(mat,radius=25,p_shared=0.8)
}

}
