% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CasPhy_functions.R
\name{splt.summerize}
\alias{splt.summerize}
\title{calculate the number of OTUs after splitting the data.}
\usage{
splt.summerize(splt)
}
\arguments{
\item{splt}{The results from `split.data`.}
}
\value{
an dataframe containing the layer, the cluster index and the number of OTUs.
}
\description{
calculate the number of OTUs after splitting the data.
}
\examples{
\dontrun{
head(splt.summerize(splt),n=20)
}

}
