% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CasPhy_functions.R
\name{to.bit}
\alias{to.bit}
\title{turn an integer matrix to a bit-coded matrix}
\usage{
to.bit(mat, check = F)
}
\arguments{
\item{mat}{The mutation matrix containing only 0 and 1.}

\item{check}{Whether to remove columns without information.}
}
\value{
an bit-coded matrix.
}
\description{
turn an integer matrix to a bit-coded matrix
}
\examples{
\dontrun{
mat<-to.bit(mat)
}

}
