% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CasPhy_functions.R
\name{cal.nevo}
\alias{cal.nevo}
\title{calculate the number of evolution events of the phylogeny.}
\usage{
cal.nevo(phy, mat, bit_coded = F)
}
\arguments{
\item{phy}{The phylogeny object.}

\item{mat}{The mutation matrix.}

\item{bit_coded}{Whether the matrix is bit coded.}
}
\value{
an bit-coded matrix marking the evolution events in the nodes of the phylogeny.
}
\description{
calculate the number of evolution events of the phylogeny.
}
\examples{
\dontrun{
nEvo<-cal.nevo(tre,mat)
}

}
