% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CasPhy_functions.R
\name{collapse.infoless}
\alias{collapse.infoless}
\title{add the length of edges to the phylogeny and collapse nodes without evolution.}
\usage{
collapse.infoless(phy, mat, bit_coded = F, collapse = T)
}
\arguments{
\item{phy}{The phylogeny object.}

\item{mat}{The mutation matrix.}

\item{bit_coded}{Whether the matrix is bit coded.}

\item{collapse}{Whether to collapse the nodes. If set to False, the function just add the length of edges to the phylogeny.}
}
\value{
a phylogeny.
}
\description{
add the length of edges to the phylogeny and collapse nodes without evolution.
}
\examples{
\dontrun{
tre<-collapse.infoless(tre,mat)
}

}
