% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CasPhy_functions.R
\name{find.dup}
\alias{find.dup}
\title{find duplicated OTUs in the mutation matrix.}
\usage{
find.dup(mat, bit_coded = F, ncore = 1)
}
\arguments{
\item{mat}{The mutation matrix.}

\item{ncore}{The number of threads used.}
}
\value{
a integer vector. The value of ith element is the index of OTU shared the same mutation profiles with the ith OTU.
}
\description{
find duplicated OTUs in the mutation matrix.
}
\examples{
\dontrun{
dup_idx<-find.dup(mat,ncore=10)
}

}
