% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CasPhy_functions.R
\name{h.partial.join}
\alias{h.partial.join}
\title{reconstruct the phylogenies using partial join algorithm}
\usage{
h.partial.join(
  mat,
  splt,
  bit_coded = F,
  keep_cutoff = 1,
  weight = NULL,
  prefix = "./",
  cuda_path = NULL,
  cuda_threshold = 500,
  ncore = 1,
  verbose = T
)
}
\arguments{
\item{mat}{The mutation matrix.}

\item{splt}{Split of data from `split.data`.}

\item{bit_coded}{Whether the matrix is bit coded.}

\item{keep_cutoff}{The cutoff to break problematic clades.}

\item{weight}{Weights of indels.}

\item{prefix}{The prefix of temporary file.}

\item{cuda_path}{The path of CUDA script. If set to `NULL`, do not use CUDA.}

\item{cuda_threshold}{Cluster larger than this value will be accelerated by CUDA.}

\item{ncore}{The number of cores used.}

\item{verbose}{Whether print detailed information.}
}
\value{
a list of phylogenies.
}
\description{
reconstruct the phylogenies using partial join algorithm
}
\examples{
\dontrun{
phys<-h.partial.join(imputed,splt,noise=F,verbose=F)
}

}
