% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CasPhy_functions.R
\name{link.hcluster}
\alias{link.hcluster}
\title{link clusters to a single phylogeny}
\usage{
link.hcluster(mg)
}
\arguments{
\item{mg}{List of phylogenies.}
}
\value{
a phylogeny.
}
\description{
link clusters to a single phylogeny
}
\examples{
\dontrun{
tre<-as.ultrametric(phys)
}

}
