% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CasPhy_functions.R
\name{shared.mutation.join}
\alias{shared.mutation.join}
\title{phylogeny reconstruction using shared-mutation-join algorithm}
\usage{
shared.mutation.join(mat, bit_coded = F, weight = NULL)
}
\arguments{
\item{mat}{The mutation matrix.}

\item{bit_coded}{Whether the matrix is bit coded.}

\item{weight}{Weights of indels.}
}
\value{
a reconstructed phylogeny.
}
\description{
phylogeny reconstruction using shared-mutation-join algorithm
}
\examples{
\dontrun{
tre<-shared.mutation.join(mat)
}

}
